+++ src/components/VisualOptionsItem.js
@@ -1,7 +1,12 @@
 import React from "react"
 import styles from "./VisualOptions.css"
+import { createBuilderStatusReporter } from "typescript"
 
 class VisualOptionsItem extends React.Component {
+  constructor(props) {
+    super(props)
+    this.state = { over: false }
+  }
   focus() {
     this.inputElement ? this.inputElement.focus() : null
   }
@@ -12,13 +17,19 @@ class VisualOptionsItem extends React.Component {
       name,
       value,
       selected,
+      showLabel = true,
+      showLabelAsTooltip = false,
       onChange,
       icon: Icon,
     } = this.props
 
     return (
       <div
-        className={`${styles.item} ${selected ? styles.selected : ""}`}
+        className={`${styles.item} ${showLabelAsTooltip ? styles.tooltip : ""} ${selected ? styles.selected : ""} ${
+          this.state.over ? styles.over : ""
+        }`}
+        onMouseOver={() => this.setState({ over: true })}
+        onMouseOut={() => this.setState({ over: false })}
         onClick={e => {
           this.inputElement.click()
           e.preventDefault()
@@ -28,7 +39,7 @@ class VisualOptionsItem extends React.Component {
         <div className={styles.icon}>
           <Icon />
         </div>
-        {name && <small>{name}</small>}
+        {(showLabel || showLabelAsTooltip) && name && <div className={styles.tip}><small>{name}</small></div>}
         <input
           type="radio"
           name={fieldName}