+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 Notable changes are within this log. Not all patch releases will be detailed, see commit history for a greater insight.
 
+## 0.8.0 - 4th May 2020
+
+### Features 
+
+- Labels can be shown as tool tips on hover by adding `showTooltip: true` to the options.
+- Option sizes can be set to `small` and `large` by adding `optionSize: "small"` to the options ([Issue #2](https://github.com/fractaldimensions/sanity-plugin-visual-options/issues/2))
+
 ## 0.7.0 - 9th April 2020
 
 Last of stability issues resolved.